sw/scripts/verifsdk/requirements.txt
Jinja2==3.1.2
pydantic==2.6.4
pydantic_core==2.16.3
pydantic-settings==2.2.1
PyYAML==6.0
